// src/app/Interfaces/iplanet.ts
import {Stringoridvalue} from '../Types/stringoridvalue';

export interface IPlanet {
	climate: string;
	created: string;
	diameter: string;
	edited: string;
	films: Stringoridvalue[];
	gravity: string;
	name: string;
	orbital_period: string;
	population: string;
	residents: Stringoridvalue[];
	rotation_period: string;
	surface_water: string;
	terrain: string;
	url: string;
}